+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## [Unreleased]
 
+### Changed
+
+#### RemoteConfig
+
+* Introduced `Kreait\Firebase\RemoteConfig\ParameterValue` to be used instead of the explicit classes 
+  `Kreait\Firebase\RemoteConfig\ExplicitValue` and `Kreait\Firebase\RemoteConfig\DefaultValue`
+  * `Kreait\Firebase\RemoteConfig\DefaultValue` should be regarded as deprecated, it is kept to not
+    create a breaking change
+  * `Kreait\Firebase\RemoteConfig\ExplicitValue` is deprecated
+
 ## [7.3.1] - 2023-06-10
 
 ### Changed